# Backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentación",
    "description": "Documentación de la API generada automáticamente",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Ingrese su token en el formato: Bearer <token>"
    }
  },
  "paths": {
...
    "/usuarios/register": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Registra un nuevo usuario",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Registro de nuevo usuario.",
            "schema": {
              "$ref": "#/definitions/RegisterUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Inicia sesión",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciales de usuario.",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },"/usuarios/actualizar/{id}": {
  "put": {
    "tags": [
      "Usuarios"
    ],
    "summary": "Actualiza un usuario por ID",
    "description": "",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "type": "integer",
        "description": "ID del usuario"
      },
      {
        "name": "authorization",
        "in": "header",
        "type": "string"
      },
      {
        "name": "body",
        "in": "body",
        "description": "Datos actualizados del usuario.",
        "schema": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "description": "Nuevo correo electrónico del usuario"
            },
            "password": {
              "type": "string",
              "description": "Nueva contraseña del usuario"
            },
            "telefono": {
              "type": "string",
              "description": "Nuevo número de teléfono del usuario"
            }
          }
        }
      }
    ],
    "responses": {
      "200": {
        "description": "OK. Usuario actualizado correctamente."
      },
      "400": {
        "description": "Error de solicitud. Verifica los datos enviados."
      },
      "404": {
        "description": "Usuario no encontrado."
      },
      "500": {
        "description": "Error interno del servidor."
      }
    },
    "security": [
      {
        "BearerAuth": []
      }
    ]
  }
},
    "/usuarios/change-password": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Cambia la contraseña del usuario",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Cambio de contraseña de usuario.",
            "schema": {
              "$ref": "#/definitions/ChangePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuarios/users": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar todos los usuarios",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtiene un usuario por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del usuario"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/usuarios/email/{email}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtiene un usuario por email",
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Email del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/usuarios/forgot-password": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Recupera la contraseña olvidada",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Recuperación de contraseña de usuario.",
            "schema": {
              "$ref": "#/definitions/ForgotPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuarios/enable-user": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Habilita un usuario",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Habilitación de usuario.",
            "schema": {
              "$ref": "#/definitions/EnableDisableUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuarios/disable-user": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Deshabilita un usuario",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Deshabilitación de usuario.",
            "schema": {
              "$ref": "#/definitions/EnableDisableUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "RegisterUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "user@example.com"
        },
        "password": {
          "type": "string",
          "example": "password123"
        },
        "role": {
          "type": "string",
          "example": "USER"
        },
        "telefono": {
          "type": "string",
          "example": "123456789"
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "user@example.com"
        },
        "password": {
          "type": "string",
          "example": "password123"
        }
      }
    },
    "ChangePassword": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "oldPassword": {
          "type": "string",
          "example": "password123"
        },
        "newPassword": {
          "type": "string",
          "example": "newpassword123"
        }
      }
    },
    "ForgotPassword": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "user@example.com"
        }
      }
    },
    "EnableDisableUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        }
      }
    },
...
  }
}